graph-models/pyproject.toml
[tool.poetry]
name = "pinochet-graph-models"
version = "0.1.0"
description = "utilities to model dataset as a graph with networkx"
authors = ["Diego Quintana <daquintanav@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
networkx = "^3.4.2"
pandas = "^2.2.3"
python-decouple = "^3.8"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
pyarrow = "^17.0.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.29.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
